# schememitra/packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SchemeMitra Core API",
    "description": "Deterministic eligibility, repayment maths, partner routing, documents, voice and application tracking for marginalized entrepreneurs. Money is integer paise; rates are bps.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/partners/nearby": {
      "get": {
        "tags": [
          "partners"
        ],
        "summary": "Nearby",
        "operationId": "nearby_v1_partners_nearby_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Lat"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Lng"
            }
          },
          {
            "name": "scheme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 48
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scheme"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 24
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 300,
              "exclusiveMinimum": 0,
              "default": 40,
              "title": "Radius Km"
            }
          },
          {
            "name": "loan_paise",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Loan Paise"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearbyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/geo/pincode/{pincode}": {
      "get": {
        "tags": [
          "partners"
        ],
        "summary": "Locate Pincode",
        "operationId": "locate_pincode_v1_geo_pincode__pincode__get",
        "parameters": [
          {
            "name": "pincode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{6}$",
              "title": "Pincode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/geo/districts": {
      "get": {
        "tags": [
          "partners"
        ],
        "summary": "Districts",
        "operationId": "districts_v1_geo_districts_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DistrictOut"
                  },
                  "title": "Response Districts V1 Geo Districts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/partner/queue": {
      "get": {
        "tags": [
          "partner"
        ],
        "summary": "Queue",
        "description": "Files sent to this lender. Admins see all lenders (read-only for decisions).",
        "operationId": "queue_v1_partner_queue_get",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "new",
                "in_review",
                "decided"
              ],
              "type": "string",
              "default": "all",
              "title": "Group"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/partner/scan": {
      "post": {
        "tags": [
          "partner"
        ],
        "summary": "Scan",
        "description": "Verify the signed QR from the citizen's phone or loan file, then open that file.",
        "operationId": "scan_v1_partner_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/partner/applications/{application_id}": {
      "get": {
        "tags": [
          "partner"
        ],
        "summary": "Review",
        "operationId": "review_v1_partner_applications__application_id__get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/partner/applications/{application_id}/decision": {
      "post": {
        "tags": [
          "partner"
        ],
        "summary": "Decision",
        "description": "receive · start_review · request_documents · approve (sanction letter) · reject (coded reason) · disburse.\nStatus push and SMS in the applicant's language go out when this commits (claim C21: the lender decides).",
        "operationId": "decision_v1_partner_applications__application_id__decision_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuditEntry": {
        "properties": {
          "at": {
            "type": "string",
            "title": "At"
          },
          "actor_role": {
            "type": "string",
            "title": "Actor Role"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "diff": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diff"
          },
          "hash": {
            "type": "string",
            "title": "Hash"
          }
        },
        "type": "object",
        "required": [
          "at",
          "actor_role",
          "action",
          "diff",
          "hash"
        ],
        "title": "AuditEntry"
      },
...
      "DecisionIn": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "receive",
              "start_review",
              "request_documents",
              "approve",
              "reject",
              "disburse"
            ],
            "title": "Action"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Documents",
            "default": []
          },
          "sanction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SanctionTerms"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "DecisionIn"
      },
...
      "DistrictOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state_code": {
            "type": "string",
            "title": "State Code"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "pincode": {
            "type": "string",
            "title": "Pincode"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "state_code",
          "lat",
          "lng",
          "pincode"
        ],
        "title": "DistrictOut"
      },
...
      "ExcludedPartner": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "state_code": {
            "type": "string",
            "title": "State Code"
          },
          "district_code": {
            "type": "string",
            "title": "District Code"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "documents_needed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Documents Needed"
          },
          "open_hours": {
            "type": "string",
            "title": "Open Hours"
          },
          "contact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Contact"
          },
          "schemes_supported": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Schemes Supported"
          },
          "categories_supported": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories Supported"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "health_score": {
            "type": "integer",
            "title": "Health Score"
          },
          "health_components": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Health Components"
          },
          "avg_sanction_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Sanction Days"
          },
          "rank_score": {
            "type": "number",
            "title": "Rank Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "state_code",
          "district_code",
          "district_name",
          "address",
          "lat",
          "lng",
          "distance_km",
          "languages",
          "documents_needed",
          "open_hours",
          "contact",
          "schemes_supported",
          "categories_supported",
          "is_demo",
          "health_score",
          "health_components",
          "avg_sanction_days",
          "rank_score",
          "reason"
        ],
        "title": "ExcludedPartner"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NearbyResponse": {
        "properties": {
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "center": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Center"
          },
          "radius_km": {
            "type": "number",
            "title": "Radius Km"
          },
          "partners": {
            "items": {
              "$ref": "#/components/schemas/PartnerCard"
            },
            "type": "array",
            "title": "Partners"
          },
          "excluded": {
            "items": {
              "$ref": "#/components/schemas/ExcludedPartner"
            },
            "type": "array",
            "title": "Excluded"
          },
          "cache": {
            "type": "string",
            "title": "Cache"
          },
          "took_ms": {
            "type": "number",
            "title": "Took Ms"
          }
        },
        "type": "object",
        "required": [
          "engine",
          "center",
          "radius_km",
          "partners",
          "excluded",
          "cache",
          "took_ms"
        ],
        "title": "NearbyResponse"
      },
...
      "PartnerCard": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "state_code": {
            "type": "string",
            "title": "State Code"
          },
          "district_code": {
            "type": "string",
            "title": "District Code"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "documents_needed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Documents Needed"
          },
          "open_hours": {
            "type": "string",
            "title": "Open Hours"
          },
          "contact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Contact"
          },
          "schemes_supported": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Schemes Supported"
          },
          "categories_supported": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories Supported"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "health_score": {
            "type": "integer",
            "title": "Health Score"
          },
          "health_components": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Health Components"
          },
          "avg_sanction_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Sanction Days"
          },
          "rank_score": {
            "type": "number",
            "title": "Rank Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "state_code",
          "district_code",
          "district_name",
          "address",
          "lat",
          "lng",
          "distance_km",
          "languages",
          "documents_needed",
          "open_hours",
          "contact",
          "schemes_supported",
          "categories_supported",
          "is_demo",
          "health_score",
          "health_components",
          "avg_sanction_days",
          "rank_score"
        ],
        "title": "PartnerCard"
      },
...
      "QueueOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "type": "array",
            "title": "Items"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "partner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Name"
          }
        },
        "type": "object",
        "required": [
          "items",
          "counts",
          "partner_name"
        ],
        "title": "QueueOut"
      },
...
      "ReviewDocument": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "sandbox": {
            "type": "boolean",
            "title": "Sandbox"
          },
          "fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Fields"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          },
          "has_image": {
            "type": "boolean",
            "title": "Has Image"
          },
          "name_match": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Match"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "type",
          "source",
          "status",
          "verified",
          "confidence",
          "sandbox",
          "fields",
          "issues",
          "has_image"
        ],
        "title": "ReviewDocument"
      },
      "ReviewOut": {
        "properties": {
          "application": {
            "$ref": "#/components/schemas/ApplicationOut"
          },
          "applicant": {
            "additionalProperties": true,
            "type": "object",
            "title": "Applicant"
          },
          "scheme": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scheme"
          },
          "finance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Finance"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/ReviewDocument"
            },
            "type": "array",
            "title": "Documents"
          },
          "readiness": {
            "additionalProperties": true,
            "type": "object",
            "title": "Readiness"
          },
          "consent": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent"
          },
          "audit": {
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            },
            "type": "array",
            "title": "Audit"
          },
          "allowed_actions": {
            "items": {
              "type": "string",
              "enum": [
                "receive",
                "start_review",
                "request_documents",
                "approve",
                "reject",
                "disburse"
              ]
            },
            "type": "array",
            "title": "Allowed Actions"
          },
          "reject_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reject Reasons"
          },
          "integrity": {
            "additionalProperties": true,
            "type": "object",
            "title": "Integrity"
          },
          "sanction_letter_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sanction Letter Url"
          }
        },
        "type": "object",
        "required": [
          "application",
          "applicant",
          "scheme",
          "finance",
          "documents",
          "readiness",
          "consent",
          "audit",
          "allowed_actions",
          "reject_reasons",
          "integrity",
          "sanction_letter_url"
        ],
        "title": "ReviewOut"
      },
...
      "ScanIn": {
        "properties": {
          "jws": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 20,
            "title": "Jws"
          }
        },
        "type": "object",
        "required": [
          "jws"
        ],
        "title": "ScanIn"
      },
      "ScanOut": {
        "properties": {
          "application_id": {
            "type": "string",
            "title": "Application Id"
          },
          "tracking_id": {
            "type": "string",
            "title": "Tracking Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "signature_valid": {
            "type": "boolean",
            "title": "Signature Valid"
          },
          "file_unchanged": {
            "type": "boolean",
            "title": "File Unchanged"
          },
          "issued_at": {
            "type": "string",
            "title": "Issued At"
          },
          "received_now": {
            "type": "boolean",
            "title": "Received Now"
          }
        },
        "type": "object",
        "required": [
          "application_id",
          "tracking_id",
          "status",
          "signature_valid",
          "file_unchanged",
          "issued_at",
          "received_now"
        ],
        "title": "ScanOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}